forms crud y cambios varios
- se cambio el tipo en materia.hsSemanales de float a int y se modifico en la base de datos también
- cruds de especialidad y materia terminados
- modificados servicios/controladores para algunas funcionalidades necesarias para los crud hechos

+++ Academia/Entidades/Materia.cs
@@ -11,7 +11,7 @@ public partial class Materia
 
     public string Descripcion { get; set; } = null!;
 
-    public float HsSemanales { get; set; }
+    public int HsSemanales { get; set; }
 
     public int HsTotales { get; set; }
